% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibrierscore.R
\name{brierProb}
\alias{brierProb}
\title{Generate probabilities for TDC brier score}
\usage{
brierProb(
  mod,
  newdata,
  time.eval,
  idvar,
  ...,
  coefficients = mod$coefficientsd,
  method = mod$method
)
}
\description{
Generate probabilities for TDC brier score
}
\keyword{internal}
