% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_satpred.R
\name{coxph.satpred}
\alias{coxph.satpred}
\title{Implemented packages}
\usage{
coxph.satpred(
  formula = NULL,
  train_df = NULL,
  test_df = NULL,
  param_grid = NULL,
  method = c("efron", "breslow", "exact"),
  finalmod = FALSE,
  ...
)
}
\description{
Implemented packages
}
