% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepsurv_satpred.R
\name{deepsurv.satpred}
\alias{deepsurv.satpred}
\title{Implemented packages}
\usage{
deepsurv.satpred(
  formula = NULL,
  train_df = NULL,
  test_df = NULL,
  num_nodes = list(h1 = c(32, 32)),
  param_grid = NULL,
  activation = "relu",
  dropout = 0.1,
  learning_rate = 1e-05,
  epochs = 10,
  batch_norm = TRUE,
  frac = 0,
  early_stopping = TRUE,
  finalmod = FALSE,
  seed = 9999,
  ...
)
}
\description{
Implemented packages
}
