% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibrierscore.R
\name{.shatfunc2}
\alias{.shatfunc2}
\title{Helper function for the TDC brier score}
\usage{
.shatfunc2(Ni, data, pred, tpnt, tau, idvar)
}
\description{
Helper function for the TDC brier score
}
\keyword{internal}
