% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_satpred.R
\name{gbm3.satpred}
\alias{gbm3.satpred}
\title{Implemented packages}
\usage{
gbm3.satpred(
  formula = NULL,
  train_df = NULL,
  test_df = NULL,
  distribution = "coxph",
  param_grid = NULL,
  n.trees = 1000,
  interaction.depth = 1,
  n.minobsinnode = 10,
  shrinkage = 0.1,
  finalmod = FALSE,
  error.method = c("auto", "OOB", "cv", "test"),
  ...
)
}
\description{
Implemented packages
}
