% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_satpred.R
\name{gbm3basehaz.gbm}
\alias{gbm3basehaz.gbm}
\alias{gbm3basehaz}
\title{Compute survival curve and cumulative hazard from a gbm3 through gbm model}
\usage{
\method{gbm3basehaz}{gbm}(fit, centered = TRUE)
}
\description{
Compute the baseline hazard
}
