% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_satpred.R
\name{gbm3survfit.gbm}
\alias{gbm3survfit.gbm}
\alias{gbm3survfit}
\title{Compute survival curve and cumulative hazard from a gbm3 through gbm model}
\usage{
\method{gbm3survfit}{gbm}(fit, newdata, ...)
}
\description{
Compute the predicted survivor and cumulative hazard function for a penalized Cox proportional hazard model.
}
