% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_satpred.R
\name{get_avesurv.coxph}
\alias{get_avesurv.coxph}
\title{Average survival for coxph}
\usage{
\method{get_avesurv}{coxph}(object, ...)

\method{get_avesurv}{coxph}(object, ...)
}
\description{
Average survival for coxph

Average survival
}
