% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibrierscore.R
\name{get_ibsTDC}
\alias{get_ibsTDC}
\title{Compute brier or integrated brier score for a TDC survival data}
\usage{
get_ibsTDC(
  mod,
  newdata,
  idvar,
  time.eval = NULL,
  type = c("BS", "IBS", "both"),
  nreps = 50,
  prop = c(0.025, 0.5, 0.975),
  parallelize = FALSE,
  nclusters = 1,
  ...,
  coefficients = mod$coefficientsd,
  method = mod$method
)
}
\description{
Compute brier or integrated brier score for a TDC survival data
}
