% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepsurv_satpred.R
\name{get_indivsurv.deepsurv}
\alias{get_indivsurv.deepsurv}
\title{Individual survival}
\usage{
\method{get_indivsurv}{deepsurv}(object, newdata)
}
\description{
Adopted from riskRegression predictRisk.deepsurv
}
