% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocfuns.R
\name{get_pvimp}
\alias{get_pvimp}
\title{Permutation variable importance}
\usage{
get_pvimp(
  model,
  newdata,
  nrep = 20,
  parallelize = TRUE,
  nclusters = parallel::detectCores(),
  ...
)
}
\description{
Computes the relative importance based on random permutation of focal variable for various survival models.
}
