% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocfuns.R
\name{get_survconcord}
\alias{get_survconcord}
\title{Compute the concordance statistic for the pcoxtime model}
\usage{
get_survconcord(object, newdata = NULL, stats = FALSE)
}
\description{
The function computes the agreement between the observed response and the predictor.
}
