% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcauc.R
\name{get_tdcauc}
\alias{get_tdcauc}
\title{Generate (TDC) AUC based on riskstAUC from risksetROC package}
\usage{
get_tdcauc(
  mod,
  newdata,
  idvar = NULL,
  time.eval = NULL,
  pred = NULL,
  pred.type = "lp",
  method = "Cox",
  span = NULL,
  order = 1,
  window = "asymmetric",
  weight = "rescale",
  nreps = 50,
  prop = c(0.025, 0.5, 0.975),
  parallelize = FALSE,
  nclusters = 1,
  ...,
  modelname = class(mod)[1]
)
}
\description{
Generate (TDC) AUC based on riskstAUC from risksetROC package
}
