% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcroc.R
\name{get_tdcroc}
\alias{get_tdcroc}
\title{Generate AUC and ROC}
\usage{
get_tdcroc(
  mod,
  newdata,
  idvar = NULL,
  time.eval = NULL,
  pred = NULL,
  cut.values = NULL,
  method = "NNE",
  lambda = NULL,
  span = NULL,
  window = "symmetric",
  nreps = 50,
  prop = c(0.025, 0.5, 0.975),
  parallelize = FALSE,
  nclusters = 1,
  ...,
  modelname = class(mod)[1]
)
}
\description{
Generate AUC and ROC
}
