% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocfuns.R
\name{get_varimp}
\alias{get_varimp}
\title{Compute variable importance of various survival models object}
\usage{
get_varimp(
  object,
  type = c("coef", "perm", "model"),
  relative = TRUE,
  newdata,
  nrep = 20,
  modelname = "model1",
  parallelize = TRUE,
  nclusters = parallel::detectCores(),
  ...
)
}
\description{
Compute variable importance of various survival models object
}
\details{
Absolute value of the coefficients (coef) corresponding the tuned model are used \code{type = perm}. Otherwise, variable level importance is computed from the model.
}
