% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cv.R
\name{modcv}
\alias{modcv}
\title{Cross-validation}
\usage{
modcv(
  formula = formula(data),
  data = sys.parent(),
  modfun,
  nfolds = 10,
  foldids = NULL,
  parallelize = FALSE,
  nclusters = parallel::detectCores(),
  ...
)
}
\description{
Cross-validation
}
