% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_train.R
\name{modtune}
\alias{modtune}
\title{Model tuning}
\usage{
modtune(
  formula = formula(data),
  data = sys.parent(),
  modfun,
  param_grid = NULL,
  nfolds = 10,
  foldids = NULL,
  ...
)
}
\description{
Model tuning
}
