% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibrierscore.R
\name{one_ibsTDC}
\alias{one_ibsTDC}
\title{Helper function for the TDC brier score}
\usage{
one_ibsTDC(
  mod,
  newdata,
  time.eval = NULL,
  idvar,
  type = c("BS", "IBS", "both"),
  ...,
  coefficients = mod$coefficientsd,
  method = mod$method
)
}
\description{
Helper function for the TDC brier score
}
\keyword{internal}
