% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcauc.R
\name{one_tdcauc}
\alias{one_tdcauc}
\title{Generate (TDC) AUC based on riskstAUC from risksetROC package}
\usage{
one_tdcauc(
  mod,
  newdata,
  pred = NULL,
  pred.type = "lp",
  predict.time = NULL,
  method = "Cox",
  span = NULL,
  order = 1,
  window = "asymmetric",
  weight = "rescale",
  ...
)
}
\description{
Generate (TDC) AUC based on riskstAUC from risksetROC package
}
\keyword{internal}
