% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcroc.R
\name{one_tdcroc}
\alias{one_tdcroc}
\title{Generate (TDC) ROC and AUC}
\usage{
one_tdcroc(
  mod,
  newdata,
  pred = NULL,
  predict.time = NULL,
  cut.values = NULL,
  method = "NNE",
  lambda = NULL,
  span = NULL,
  window = "symmetric"
)
}
\description{
Generate (TDC) ROC and AUC
}
\keyword{internal}
