% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_satpred.R
\name{plot.coxph.satpred}
\alias{plot.coxph.satpred}
\title{Cross-validation plots}
\usage{
\method{plot}{coxph.satpred}(x, ..., show_best = TRUE, lsize = 0.3, pshape = "O")
}
\description{
Cross-validation plots
}
