% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepsurv_satpred.R
\name{plot.deepsurv.satpred}
\alias{plot.deepsurv.satpred}
\title{Cross-validation plots}
\usage{
\method{plot}{deepsurv.satpred}(x, ..., show_best = TRUE, lsize = 0.3, pshape = "O")
}
\description{
Cross-validation plots
}
