% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_satpred.R
\name{plot.gbm.satpred}
\alias{plot.gbm.satpred}
\title{Cross-validation plots}
\usage{
\method{plot}{gbm.satpred}(x, ..., show_best = TRUE, lsize = 0.3, pshape = "O")
}
\description{
Cross-validation plots
}
