% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibrierscore.R
\name{plot.ibsTDC}
\alias{plot.ibsTDC}
\title{Plot brier or integrated brier score estimates}
\usage{
\method{plot}{ibsTDC}(x, ..., which.plot = c("BS", "IBS", "both"), plotit = TRUE)
}
\description{
Plot brier or integrated brier score estimates
}
