% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satpredplots.R
\name{plot.rfsrc.satpred}
\alias{plot.rfsrc.satpred}
\title{Cross-validation plots}
\usage{
\method{plot}{rfsrc.satpred}(x, ..., show_best = TRUE, lsize = 0.3, pshape = "O")
}
\description{
Cross-validation plots
}
