% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcroc.R
\name{plot.roctdc}
\alias{plot.roctdc}
\title{Plot AUC and ROC supporting TDC}
\usage{
\method{plot}{roctdc}(
  x,
  ...,
  which.plot = c("both", "auc", "roc"),
  auc.type = c("line", "point"),
  include.null = FALSE,
  facet.roc = TRUE,
  ci = TRUE,
  pos = 0.3,
  size = 0.2
)
}
\description{
Plot AUC and ROC supporting TDC
}
