% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satpredplots.R
\name{plot.satsurv}
\alias{plot.satsurv}
\title{Plot survival and cumulative hazard curves}
\usage{
\method{plot}{satsurv}(
  x,
  ...,
  type = c("surv", "cumhaz"),
  lsize = 0.3,
  lcol = "grey",
  ltype = 2,
  compare = FALSE,
  compare_lab = "satpred"
)
}
\description{
Plot estimated survival and cumulative  hazard curves for \code{satpred} model.
}
