% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satpredplots.R
\name{plot.Score}
\alias{plot.Score}
\title{Prediction performance}
\usage{
\method{plot}{Score}(x, ..., type = c("roc", "auc", "brier"), pos = 0.3, psize = 0.3)
}
\arguments{
\item{x}{\code{\link[riskRegression]{Score}} object. See examples.}

\item{...}{for future implementations.}

\item{type}{metric to return. Choices are \code{"roc", "auc", "brier"}.}

\item{pos}{spacing between the lines.}
}
\value{
a \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plots predictive performance comparison of survival models. It uses risk scoring from \code{\link[riskRegression]{Score}}.
}
\details{
Implements plot method for \code{\link[riskRegression]{Score}} for time-dependent Brier score, AUC and ROC. However, currently, no support for time-dependent covariate models.
}
