% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcauc.R
\name{plot.tdcauc}
\alias{plot.tdcauc}
\title{Plot AUC and survival curves supporting TDC using risksetROC package}
\usage{
\method{plot}{tdcauc}(
  x,
  ...,
  which.plot = c("auc", "cindex", "surv"),
  auc.type = c("line", "point"),
  include.null = FALSE,
  ci = TRUE,
  pos = 0.3,
  size = 0.2
)
}
\description{
Plot AUC and survival curves supporting TDC using risksetROC package
}
