% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satpredplots.R
\name{plot.varimp}
\alias{plot.varimp}
\title{Generic method for plotting variable importance of various models}
\usage{
\method{plot}{varimp}(x, ..., pos = 0.5, drop_zero = TRUE)
}
\description{
Generic method for plotting variable importance of various models
}
