% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satpredplots.R
\name{plotpec}
\alias{plotpec}
\title{Plotting prediction error curves}
\usage{
plotpec(
  x,
  ...,
  lsize = 0.3,
  ltype = 2,
  xlab = "Time",
  ylab = "Prediction error"
)
}
\description{
Plotting prediction error curves
}
