% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepsurv_satpred.R
\name{predictRisk.deepsurv}
\alias{predictRisk.deepsurv}
\alias{predictRisk}
\title{Extract predictions from gbm model}
\usage{
\method{predictRisk}{deepsurv}(object, newdata, times, ...)
}
\description{
Extract event probabilities from the fitted model.
}
\details{
For survival outcome, the function predicts the risk, \eqn{1 - S(t|x)}, where \eqn{S(t|x)} is the survival chance of an individual characterized by \eqn{x}. riskRegression::predictRisk.gbm seems to have issues reconstructing the data.
}
