% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocfuns.R
\name{predictRisk.satpred}
\alias{predictRisk.satpred}
\alias{predictRisk}
\title{Extract predictions from satpred model}
\usage{
predictRisk.satpred(object, newdata, times, ...)
}
\description{
Extract event probabilities from the fitted model.
}
\details{
For survival outcome, the function predicts the risk, \eqn{1 - S(t|x)}, where \eqn{S(t|x)} is the survival chance of an individual characterized by \eqn{x}.
}
