% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepsurv_satpred.R
\name{predictSurvProb.deepsurv}
\alias{predictSurvProb.deepsurv}
\alias{predictSurvProb}
\title{Predict survival probabilities at various time points}
\usage{
predictSurvProb.deepsurv(object, newdata, times, ...)
}
\value{
a matrix of probabilities with as many rows as the rows of the \code{newdata} and as many columns as number of time points (\code{times}).
}
\description{
The function extracts the survival probability predictions from a \code{satpred} model.
}
