% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocfuns.R
\name{predictSurvProb.gbm.satpred}
\alias{predictSurvProb.gbm.satpred}
\alias{predictSurvProb}
\title{Predict survival probabilities at various time points}
\usage{
\method{predictSurvProb}{gbm.satpred}(object, newdata, times, ...)
}
\value{
a matrix of probabilities with as many rows as the rows of the \code{newdata} and as many columns as number of time points (\code{times}).
}
\description{
The function extracts the survival probability predictions from a \code{satpred} model.
}
