% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pvimp.default}
\alias{pvimp.default}
\title{Permutation variable importance method for default models}
\usage{
\method{pvimp}{default}(
  model,
  newdata,
  nrep = 20,
  parallelize = TRUE,
  nclusters = parallel::detectCores(),
  ...
)
}
\description{
Permutation variable importance method for default models
}
\keyword{internal}
