% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_satpred.R
\name{pvimp.gbm}
\alias{pvimp.gbm}
\title{Permutation variable importance method for gbm}
\usage{
\method{pvimp}{gbm}(model, newdata, nrep = 50)
}
\description{
Permutation variable importance method for gbm
}
\keyword{internal}
