% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pvimp.rfsrc}
\alias{pvimp.rfsrc}
\title{Permutation variable importance method for rfsrc}
\usage{
\method{pvimp}{rfsrc}(model, newdata, nrep = 50)
}
\description{
Permutation variable importance method for rfsrc
}
\keyword{internal}
