% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibrierscore.R
\name{resamplesTDC}
\alias{resamplesTDC}
\title{Create resample indices or data frame for a TDC survival data}
\usage{
resamplesTDC(df, nreps, idvar, return_index = TRUE, ...)
}
\description{
Create resample indices or data frame for a TDC survival data
}
