% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rfsrc.satpred}
\alias{rfsrc.satpred}
\title{Implemented packages}
\usage{
rfsrc.satpred(
  formula = NULL,
  train_df = NULL,
  test_df = NULL,
  param_grid = NULL,
  ntree = 1000,
  mtry = NULL,
  nodesize = NULL,
  splitrule = "logrank",
  forest = FALSE,
  finalmod = FALSE,
  ...
)
}
\description{
Implemented packages
}
