% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepsurv_satpred.R
\name{survconcord.deepsurv}
\alias{survconcord.deepsurv}
\title{Compute the concordance statistic for the deepsurv model}
\usage{
\method{survconcord}{deepsurv}(object, newdata = NULL, stats = FALSE, ...)
}
\description{
The function computes the agreement between the observed response and the predictor.
}
\keyword{internal}
