% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{survconcord.default}
\alias{survconcord.default}
\title{Compute the concordance statistic for the default model (coxph, glmnetsurv)}
\usage{
\method{survconcord}{default}(object, newdata = NULL, stats = FALSE, ...)
}
\description{
The function computes the agreement between the observed response and the predictor.
}
\keyword{internal}
