% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_satpred.R
\name{survconcord.gbm}
\alias{survconcord.gbm}
\title{Compute the concordance statistic for the pcoxtime model}
\usage{
\method{survconcord}{gbm}(object, newdata = NULL, stats = FALSE, ...)
}
\description{
The function computes the agreement between the observed response and the predictor.
}
\keyword{internal}
