% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{survconcord.rfsrc}
\alias{survconcord.rfsrc}
\title{Compute the concordance statistic for the pcoxtime model}
\usage{
\method{survconcord}{rfsrc}(object, newdata = NULL, stats = FALSE)
}
\description{
The function computes the agreement between the observed response and the predictor.
}
\keyword{internal}
