% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepsurv_satpred.R
\name{varimp.deepsurv}
\alias{varimp.deepsurv}
\title{Compute variable importance deepsurv}
\usage{
\method{varimp}{deepsurv}(
  object,
  type = c("coef", "perm", "model"),
  relative = TRUE,
  newdata,
  nrep = 20,
  ...
)
}
\description{
Compute variable importance deepsurv
}
\keyword{internal}
