% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{varimp.rfsrc}
\alias{varimp.rfsrc}
\title{Compute variable importance random forest}
\usage{
\method{varimp}{rfsrc}(
  object,
  type = c("coef", "perm", "model"),
  relative = TRUE,
  newdata,
  nrep = 20,
  parallelize = TRUE,
  nclusters = parallel::detectCores(),
  ...
)
}
\description{
Compute variable importance random forest
}
\keyword{internal}
