% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsExport.R
\name{pkgsExport}
\alias{pkgsExport}
\alias{Surv}
\title{Import and export various functions from various packages}
\description{
Import and export various functions from various packages

Import and export Surv. See \code{\link[survival]{Surv}}
}
\keyword{internal}
