% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepsurv_satpred.R
\name{pvimp.deepsurv}
\alias{pvimp.deepsurv}
\title{Permutation variable importance method for deepsurv}
\usage{
\method{pvimp}{deepsurv}(model, newdata, nrep = 50)
}
\description{
Permutation variable importance method for deepsurv
}
\keyword{internal}
