% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_satpred.R
\name{varimp.gbm}
\alias{varimp.gbm}
\title{Compute variable importance gbm}
\usage{
\method{varimp}{gbm}(
  object,
  type = c("coef", "perm", "model"),
  relative = TRUE,
  newdata,
  nrep = 20,
  ...
)
}
\description{
Compute variable importance gbm
}
\keyword{internal}
